requirements.txt
facebook-sdk
opencv-python
argparse
numpy
tqdm